+++ src/Headers/GoogleAnalyticsCampaignHeader.php
@@ -6,10 +6,10 @@ namespace Spatie\MailcoachMailer\Headers;
 
 use Symfony\Component\Mime\Header\UnstructuredHeader;
 
-class UtmSourceHeader extends UnstructuredHeader
+class GoogleAnalyticsCampaignHeader extends UnstructuredHeader
 {
     public function __construct(string $value)
     {
-        parent::__construct('X-Mailcoach-Utm-Source', $value);
+        parent::__construct('X-Mailcoach-Google-Analytics-Campaign', $value);
     }
 }