+++ src/MailcoachApiTransport.php
@@ -8,13 +8,11 @@ use Spatie\MailcoachMailer\Exceptions\EmailNotValid;
 use Spatie\MailcoachMailer\Exceptions\NoHostSet;
 use Spatie\MailcoachMailer\Exceptions\NotAllowedToSendMail;
 use Spatie\MailcoachMailer\Headers\FakeHeader;
+use Spatie\MailcoachMailer\Headers\GoogleAnalyticsCampaignHeader;
 use Spatie\MailcoachMailer\Headers\MailerHeader;
 use Spatie\MailcoachMailer\Headers\ReplacementHeader;
 use Spatie\MailcoachMailer\Headers\StoreContentHeader;
 use Spatie\MailcoachMailer\Headers\TransactionalMailHeader;
-use Spatie\MailcoachMailer\Headers\UtmCampaignHeader;
-use Spatie\MailcoachMailer\Headers\UtmMediumHeader;
-use Spatie\MailcoachMailer\Headers\UtmSourceHeader;
 use Symfony\Component\Mailer\Envelope;
 use Symfony\Component\Mailer\Exception\HttpTransportException;
 use Symfony\Component\Mailer\Exception\TransportException;
@@ -126,16 +124,8 @@ class MailcoachApiTransport extends AbstractApiTransport
                 $payload['store_content'] = $header->getValue();
             }
 
-            if ($header instanceof UtmSourceHeader) {
-                $payload['utm_source'] = $header->getValue();
-            }
-
-            if ($header instanceof UtmMediumHeader) {
-                $payload['utm_medium'] = $header->getValue();
-            }
-
-            if ($header instanceof UtmCampaignHeader) {
-                $payload['utm_campaign'] = $header->getValue();
+            if ($header instanceof GoogleAnalyticsCampaignHeader) {
+                $payload['google_analytics_campaign'] = $header->getValue();
             }
         }